+++ go/firebase-adminsdk.json
@@ -1,0 +1,12 @@
+{
+  "type": "service_account",
+  "project_id": "hacku-2021",
+  "private_key_id": "47f691ab07421a6748b3e19125fcbafd28921a2b",
+  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQCpTv8HHfu3UJMS\nIqVU1b5Ji4djZl+QqJOorY7GkFMa2juc3Cm3RTa9v61IWxUke+m33guHDgNFfaPl\ndk3vrsc3QCaCHukMnqGlBei3DCKAg6IjEmL9ZLydgtrlphswuZRneryDeKT0RVpl\ncBThiZ+CkjGCfwvINqfpQenCAjH0+LiA/YRNrccKrwLWKhIiWY0yP6ct0flg4WWY\nMBq4aOCURVeWE2fASiwRHhCYw5LlorEbJ/mNFF6aT+S+7W5H+rLGwwPDIYqFwRTg\ndKrPK5A7dGrR4cWMFn5FIbR4E+8rGA+n26rY8rC1O6gjn0kn397ja3i26dfHoq6j\n6neXrgXRAgMBAAECggEACeooMW9WOhtxVAq3BNtJe+lnAcamdQRc3/AIoDnRFrhN\n7Qx63ZJ44gQTWTKK7+qDusfkLpdUULb1U0QqIeLpzqNRPfsiaaHxKbFJZKLEyHgs\n6qyHzc/0OFhyiifqiArDaos7oTVTmkG6Hp5BfXL/kM1lMdT6cyc5jlrU82QJLzpj\nFyQN33wTZ27ual/vB6liGC5KHM8mCjDZWNr4l/AX7qZOoLaQqWDRiuVkxoPuR/Ts\nNdhs22nqpjMy66vH/xgh5q2hmGU/m3pHGZ84rSeM9bkBePnpIyJ6iVHgMy5g8iqe\nqkHo3pjJelVE8JxidCTP10me4nS5o1xrkzP5GvGMeQKBgQDT4giqbOa3FrH1ojr7\ndjA8bD9ncEIN8gjlSuTGRCOd4CzWavdYankf4WVTlLqHasBf1W/FfW+pf05QOt5l\nP4TlatTEcQ4vrG8Rd+pNZ8akYuA/n0zVEv2R2D5R3XDSNSzK/WjvHiSg0CSB91EF\n2Ainv932b0rmE0oiIFKLD8pt1QKBgQDMj6CGWbz2yziM/790+ZPY8i2WncKaTxzt\nGBk9BloDptbFbYMSISTM1hvQ7t19SPnDlewkMjvJ2pdP6HUdc+83J1UIE5hfysww\nLHOqXd+r+wuDDzmlnd2tv3OndhmmOVSYLTpkn+PyfhUeYGyyse8TBeIOZe4/0hHk\nHF2GqBuqDQKBgGMwdY4S3Q7o+PZcs+/0LDCdHDF4Q/PzUmbE4MRQMbw+rb7S5Gfn\nsDxUb/LotUmfl99FM8XoKtXWWvsEFAih67zafjBYHKQEvRoYYP9LjZp3DzO5h2ZY\nNYyJBuOcHqzJZ+a/KtiF6++8gNq2xJ9Puv2MvIQZJZ7/eLzxFLM0+TbhAoGBALkP\nIqEIpBv8w39Di1huxSNaDIYa2Mey1K9lwsdjbKGrGNdo+E+5bkplVKrSvcuu7k3T\n3trsGwzE1Wi0Gud3T61sxs9Ust/OXqUiQc2ZuVUj+0kox/UdeUPLVPzlqdSoPvCg\n0hfgUl65U6+0egG7WVazVsvBySUxjX/5bV75zHUtAoGAL/XlRP+lhyRRJ/vp6BT8\n1nKdCVaXhETIpIgPjmtB+EPR1uLXvK3ArK4wOUu6oN+VA3B/kAarUwsa7PFmrsc0\n9DYgTyo+dLW98G3XzIBzZj//3LZeoI0yALui/GCx9bfqq2/pFqJLqBM+7YnrK1pO\n6Rt0JLDkzwAh5519+wNSRis=\n-----END PRIVATE KEY-----\n",
+  "client_email": "firebase-adminsdk-p118a@hacku-2021.iam.gserviceaccount.com",
+  "client_id": "107844263280378319107",
+  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
+  "token_uri": "https://oauth2.googleapis.com/token",
+  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
+  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-p118a%40hacku-2021.iam.gserviceaccount.com"
+}
